Using the Array Filter Method
In this lab we created three arrow functions the first first functions takes in an array of driver names and a string as an argument and filters through the drivers and returns a matching list of drivers and whos names are lowercase. The second function takes in an array of drivers and a string as an argument and returns all drivers whos names begin with the provide first letter. The last arrow function takes in an array of driver objects and a string as an arugment and returns each element whos name property matches the provided string argument